Move console SM to metrics instrumentation
